made separate files for cudf_buffer
