+++ cpp/src/gcylon/all2all/cudf_all_to_all.cpp
@@ -70,39 +70,6 @@ cudf::size_type dataLength(cudf::column_view const& cw){
 }
 
 //////////////////////////////////////////////////////////////////////
-// CudfBuffer implementations
-//////////////////////////////////////////////////////////////////////
-CudfBuffer::CudfBuffer(std::shared_ptr<rmm::device_buffer> rmm_buf) : rmm_buf(rmm_buf) {}
-
-int64_t CudfBuffer::GetLength() {
-  return rmm_buf->size();
-}
-
-uint8_t * CudfBuffer::GetByteBuffer() {
-  return (uint8_t *)rmm_buf->data();
-}
-
-std::shared_ptr<rmm::device_buffer> CudfBuffer::getBuf() const {
-  return rmm_buf;
-}
-
-//////////////////////////////////////////////////////////////////////
-// CudfAllocator implementations
-//////////////////////////////////////////////////////////////////////
-cylon::Status CudfAllocator::Allocate(int64_t length, std::shared_ptr<cylon::Buffer> *buffer) {
-  try {
-    auto rmm_buf = std::make_shared<rmm::device_buffer>(length, rmm::cuda_stream_default);
-    *buffer = std::make_shared<CudfBuffer>(rmm_buf);
-    return cylon::Status::OK();
-  } catch (rmm::bad_alloc * badAlloc) {
-    LOG(ERROR) << "failed to allocate gpu memory with rmm: " << badAlloc->what();
-    return cylon::Status(cylon::Code::GpuMemoryError);
-  }
-}
-
-CudfAllocator::~CudfAllocator() = default;
-
-//////////////////////////////////////////////////////////////////////
 // PendingBuffer implementations
 //////////////////////////////////////////////////////////////////////
 PendingBuffer::PendingBuffer(const uint8_t *buffer,
